REFACTOR: (some tests failing) event -> abstract class to avoid duplicating method code; get circuit breaker and DER tests passing
