+++ dev-notes/019.4-ReorganizeEventBus.md
@@ -168,12 +168,18 @@ Add `publishEventsBulk` for BullMq bus
 
 **COMMIT: REFACTOR: (some tests failing) build memory event bus for new interface; change application to use new event bus interface**
 
+-  `incrementRetryCount()`, `incremementConnectFailureCount()`, `get eventType()`
+-  Change `IEventBusEvent` to `EventBusEvent`, an abstract class that implements the methods
+   -  Descendant constructors need to set `_eventData` and `_eventKey`
+   -  Avoids duplicating method code
+-  Fix circuit breaker and DER tests
+   -  Mostly changes to types due to event bus reorg
+
+**COMMIT: REFACTOR: (some tests failing) event -> abstract class to avoid duplicating method code; get circuit breaker and DER tests passing**
+
 ## TODO
 
--  Subscribers increment retryCount directly -- give them a setter; changes to all events plus interface and subscribers
--  Add `eventType` getter -- similar scope
--  Circuit breaker and DER tests need a lot of work
--  Move bullmq connection into `eventBus`
+-  Move bullmq connection into `eventBus`; add env support; what to do about BullMq type, etc.?
 -  Several tests assume BullMq
    -  Memory event bus does not await handler run
    -  BullMq bus does await because it needs to know if publish fails
