+++ src/backup-request/domain/BackupRequestAllowed.event.ts
@@ -1,4 +1,4 @@
-import { IEventBusEvent, IEventBusEventData } from '../../common/infrastructure/event-bus/IEventBus';
+import { EventBusEvent } from '../../common/infrastructure/event-bus/IEventBus';
 import { BackupRequest } from './BackupRequest';
 
 export interface BackupRequestAllowedEventData {
@@ -9,15 +9,10 @@ export interface BackupRequestAllowedEventData {
 	};
 }
 
-export class BackupRequestAllowed implements IEventBusEvent {
-	private _eventTimestamp: Date;
-	private _eventData: IEventBusEventData & BackupRequestAllowedEventData;
-	private _eventKey: string;
-	private _topicName: string;
+export class BackupRequestAllowed extends EventBusEvent<BackupRequestAllowedEventData> {
 
 	constructor(backupRequest: BackupRequest) {
-		this._eventTimestamp = new Date();
-		this._topicName = 'linebacker';
+		super();
 		this._eventData = {
 			connectFailureCount: 0,
 			retryCount: 0,
@@ -27,41 +22,5 @@ export class BackupRequestAllowed implements IEventBusEvent {
 			},
 		};
 		this._eventKey = this._eventData.event.backupRequestId;
-	}
-
-	get topicName() {
-		return this._topicName;
-	}
-
-	get eventKey() {
-		return this._eventKey;
-	}
-
-	get eventData() {
-		return this._eventData;
-	}
-
-	get event() {
-		return this._eventData.event;
-	}
-
-	get eventDataString() {
-		return JSON.stringify(this._eventData);
-	}
-
-	get eventTimestamp() {
-		return this._eventTimestamp;
-	}
-
-	get retryCount() {
-		return this._eventData.retryCount;
-	}
-
-	get connectFailureCount() {
-		return this._eventData.connectFailureCount;
-	}
-
-	incrementRetryCount() {
-		this._eventData.retryCount++;
 	}
 }
