+++ src/backup-request/domain/StoreStatusReceived.event.ts
@@ -1,5 +1,5 @@
 
-import { IEventBusEvent, IEventBusEventData } from '../../common/infrastructure/event-bus/IEventBus';
+import { EventBusEvent } from '../../common/infrastructure/event-bus/IEventBus';
 
 export interface StoreStatusReceivedEventData {
 	event: StoreStatusMessageItem;
@@ -26,15 +26,10 @@ export interface StoreStatusMessage {
 	messageText?: string;
 }
 
-export class StoreStatusReceived implements IEventBusEvent {
-	private _eventTimestamp: Date;
-	private _eventData: IEventBusEventData & StoreStatusReceivedEventData;
-	private _eventKey: string;
-	private _topicName: string;
+export class StoreStatusReceived extends EventBusEvent<StoreStatusReceivedEventData> {
 
 	constructor(queueMessage: StoreStatusMessageItem) {
-		this._eventTimestamp = new Date();
-		this._topicName = 'linebacker';
+		super();
       const { dequeueCount, messageId, popReceipt, messageObject } = queueMessage;
 		this._eventData = {
 			connectFailureCount: 0,
@@ -60,41 +55,5 @@ export class StoreStatusReceived implements IEventBusEvent {
 			},
 		};
 		this._eventKey = `${this._eventData.event.messageId}|${this._eventData.event.popReceipt}`;
-	}
-
-	get topicName() {
-		return this._topicName;
-	}
-
-	get eventKey() {
-		return this._eventKey;
-	}
-
-	get eventData() {
-		return this._eventData;
-	}
-
-	get event() {
-		return this._eventData.event;
-	}
-
-	get eventDataString() {
-		return JSON.stringify(this._eventData);
-	}
-
-	get eventTimestamp() {
-		return this._eventTimestamp;
-	}
-
-	get retryCount() {
-		return this._eventData.retryCount;
-	}
-
-	get connectFailureCount() {
-		return this._eventData.connectFailureCount;
-	}
-
-	incrementRetryCount() {
-		this._eventData.retryCount++;
 	}
 }
