+++ src/common/infrastructure/event-bus/MemoryEventBus.ts
@@ -2,7 +2,7 @@ import path from 'node:path';
 import { logger } from '../../../infrastructure/logging/pinoLogger';
 import { ok, Result } from '../../core/Result';
 import { EventBusError } from '../InfrastructureErrors';
-import { EventBusHandler, IEventBus, IEventBusEvent } from './IEventBus';
+import { EventBusHandler, IEventBus, EventBusEvent } from './IEventBus';
 
 const moduleName = path.basename(module.filename);
 
@@ -13,7 +13,7 @@ export class MemoryEventBus implements IEventBus {
         this._eventHandlers = new Map<string, EventBusHandler[]>;
     }
 
-    async publishEvent(event: IEventBusEvent): Promise<Result<IEventBusEvent, EventBusError>> {
+    async publishEvent(event: EventBusEvent<unknown>): Promise<Result<EventBusEvent<unknown>, EventBusError>> {
         const functionName = 'publishEvent';
         const eventName: string = event.constructor.name;
 
@@ -27,7 +27,7 @@ export class MemoryEventBus implements IEventBus {
         return ok(event);
     }
 
-    async publishEventsBulk(events: IEventBusEvent[]): Promise<Result<IEventBusEvent[], EventBusError>> {
+    async publishEventsBulk(events: EventBusEvent<unknown>[]): Promise<Result<EventBusEvent<unknown>[], EventBusError>> {
 		for (const ev of events) {
 			await this.publishEvent(ev);
 		};
@@ -35,7 +35,7 @@ export class MemoryEventBus implements IEventBus {
 		return ok(events);
     }
 
-    subscribe(eventName: string, handler: (event: IEventBusEvent) => void): void {
+    subscribe(eventName: string, handler: (event: EventBusEvent<unknown>) => void): void {
         // either the array for the event's handlers or an empty array
         const eventHandlers = this._eventHandlers.get(eventName) || [];
         // only add handler if it isn't already registered
