+++ src/infrastructure/resilience/ApplicationResilienceReady.event.ts
@@ -1,4 +1,4 @@
-import { IEventBusEvent, IEventBusEventData } from '../../common/infrastructure/event-bus/IEventBus';
+import { EventBusEvent } from '../../common/infrastructure/event-bus/IEventBus';
 
 export interface ApplicationResilienceReadyEventData {
 	// After received, all consumers must get data from persisted data
@@ -8,15 +8,10 @@ export interface ApplicationResilienceReadyEventData {
 	};
 }
 
-export class ApplicationResilienceReady implements IEventBusEvent {
-	private _eventTimestamp: Date;
-	private _eventData: IEventBusEventData & ApplicationResilienceReadyEventData;
-	private _eventKey: string;
-	private _topicName: string;
+export class ApplicationResilienceReady extends EventBusEvent<ApplicationResilienceReadyEventData> {
 
 	constructor(beforeTimestamp: Date) {
-		this._eventTimestamp = new Date();
-		this._topicName = 'linebacker';
+		super();
 		this._eventData = {
 			connectFailureCount: 0,
 			retryCount: 0,
@@ -26,41 +21,5 @@ export class ApplicationResilienceReady implements IEventBusEvent {
 			},
 		};
 		this._eventKey = this._eventData.event.beforeTimestamp.toISOString();
-	}
-
-	get topicName() {
-		return this._topicName;
-	}
-
-	get eventKey() {
-		return this._eventKey;
-	}
-
-	get eventData() {
-		return this._eventData;
-	}
-
-	get event() {
-		return this._eventData.event;
-	}
-
-	get eventDataString() {
-		return JSON.stringify(this._eventData);
-	}
-
-	get eventTimestamp() {
-		return this._eventTimestamp;
-	}
-
-	get retryCount() {
-		return this._eventData.retryCount;
-	}
-
-	get connectFailureCount() {
-		return this._eventData.connectFailureCount;
-	}
-
-	incrementRetryCount() {
-		this._eventData.retryCount++;
 	}
 }
