+++ src/infrastructure/resilience/DelayedEventRunner.ts
@@ -1,5 +1,5 @@
 import { eventBus } from '../../common/infrastructure/event-bus/eventBus';
-import { IEventBusEvent } from '../../common/infrastructure/event-bus/IEventBus';
+import { EventBusEvent } from '../../common/infrastructure/event-bus/IEventBus';
 import { delay } from '../../common/utils/utils';
 
 const DelayedEventRunnerStateValues = {
@@ -11,7 +11,7 @@ const DelayedEventRunnerStateValues = {
 type DelayedEventRunnerStateType = typeof DelayedEventRunnerStateValues[keyof typeof DelayedEventRunnerStateValues];
 
 export class DelayedEventRunner {
-	private _events: IEventBusEvent[] = [];
+	private _events: EventBusEvent<any>[] = [];
 	private _state: DelayedEventRunnerStateType = DelayedEventRunnerStateValues.Stop;
 	private _delayMs: number;
 	private _abortSignal: AbortSignal;
@@ -27,7 +27,7 @@ export class DelayedEventRunner {
 
 	public get events() {
 		return this._events.map((ev) => {
-			return { ...ev };
+			return ev;
 		});
 	}
 
@@ -63,7 +63,7 @@ export class DelayedEventRunner {
 		this._state = DelayedEventRunnerStateValues.Stop;
 	}
 
-	public addEvent(ev: IEventBusEvent): void {
+	public addEvent(ev: EventBusEvent<unknown>): void {
 		if (this.isStateHalt()) return;
 
 		// only add when ev doesn't already exist in event array
@@ -91,7 +91,7 @@ export class DelayedEventRunner {
 		// console.log('DER run events', this.events);
 		this.setStateRun();
 
-		let ev: IEventBusEvent | undefined;
+		let ev: EventBusEvent<unknown> | undefined;
 		// Array.prototype.shift() removes the first element from the array and returns it or undefined if none
 		// order of this condition is important to ensure we don't shift events off the array if we're stopping
 		while (this.isStateRun() && typeof (ev = this._events.shift()) !== 'undefined') {
